backend/ml_model/requirements.txt
faiss-cpu
sentence-transformers
scikit-learn
numpy
pandas
flask
transformers
torch
